+++ internal/interfaces/storage_handler.go
@@ -2,6 +2,6 @@ package interfaces
 
 type StorageHandler interface {
 	Key(...string) string
-	Get(string) (*uint64, error)
-	Set(string, uint64) (*uint64, error)
+	Get(string) (*float64, error)
+	Set(string, float64) (*float64, error)
 }
